+++ applications/services/cli/cli_command_sl_cli.c
@@ -75,3 +75,53 @@ void cli_command_sl_cli(Cli* cli, FuriString* args, void* context) {
 
     cli_command_sl_cli_free(instance);
 }
+
+static void cli_send_sl_cli_cmd(FuriString* cmd) {
+    Intercom* intercom = furi_record_open(RECORD_INTERCOM);
+    size_t sz = furi_string_size(cmd);
+    size_t tx_size =
+        intercom_tx(intercom, IntercomChannelCli, furi_string_get_cstr(cmd), sz, FuriWaitForever);
+    furi_assert(tx_size == sz);
+    furi_record_close(RECORD_INTERCOM);
+}
+
+void cli_command_sl_echo(Cli* cli, FuriString* args, void* context) {
+    UNUSED(context);
+    UNUSED(args);
+
+    const uint32_t baud = 230400UL;
+
+    printf("Starting 917 echo server on %ld\r\n", baud);
+    FuriString* cmd = furi_string_alloc_printf("echo_server  %ld\r", baud);
+    cli_send_sl_cli_cmd(cmd);
+
+    FuriHalSerialHandle* serial = furi_hal_serial_control_acquire(FuriHalSerialIdUsart6);
+    furi_hal_serial_init(serial, baud);
+    furi_hal_serial_clear(serial, FuriHalSerialDirectionTxRx);
+
+    while(true) {
+        uint8_t ch = cli_getc(cli);
+
+        if(ch == CliSymbolAsciiETX) {
+            break;
+        } else if(ch == CliSymbolAsciiCR || ch == CliSymbolAsciiLF)
+            continue;
+
+        furi_hal_serial_tx(serial, &ch, 1);
+        if(!furi_hal_serial_tx_wait_complete(serial, 100)) {
+            break;
+        }
+
+        furi_delay_ms(10);
+
+        while(furi_hal_serial_rx_available(serial)) {
+            ch = furi_hal_serial_rx(serial);
+            cli_putc(cli, ch);
+        }
+    }
+
+    furi_hal_serial_control_release(serial);
+    furi_string_printf(cmd, "%c\r", CliSymbolAsciiETX);
+    cli_send_sl_cli_cmd(cmd);
+    furi_string_free(cmd);
+}
