+++ applications/services/cli/cli_command_sl_cli.h
@@ -3,3 +3,4 @@
 #include "cli_i.h"
 
 void cli_command_sl_cli(Cli* cli, FuriString* args, void* context);
+void cli_command_sl_echo(Cli* cli, FuriString* args, void* context);
